+++ src/index.js
@@ -20,7 +20,7 @@ import CustomStateResults from './components/CustomStateResults';
 import carousel from './styles/carousel';
 import Carousel from './components/Carousels/Carousel';
 import CardCarousel from './components/Carousels/CardCarousel';
-import CardWrapper from './components/Cards/CardWrapper';
+import CardWrapper from './components/Cards/CardWrapper/CardWrapper';
 import CategoryCard from './components/Cards/CategoryCard';
 import ClearRefinements from './components/Algolia/shared/ClearRefinements';
 import CookingSchoolAd from './components/Ads/ReviewsAds/CookingSchoolAd';
